implement tests for auth

+++ services/garage-backend/tests/db.ts
@@ -20,10 +20,12 @@ const closeDatabase = async () => {
 const clearDatabase = async () => {
   const collections = mongoose.connection.collections;
 
-  for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany({});
-  }
+  await Promise.all(
+    Object.values(collections).map((collection) => {
+      if (collection.name === 'roles') return;
+      collection.deleteMany({});
+    }),
+  );
 };
 
 export default {
